+++ src/main/java/site/sonisori/sonisori/auth/oauth2/dto/NaverResponse.java
@@ -6,7 +6,7 @@ public class NaverResponse implements OAuth2Response {
 	private final Map<String, Object> attribute;
 
 	public NaverResponse(Map<String, Object> attribute) {
-		this.attribute = attribute;
+		this.attribute = (Map<String, Object>)attribute.get("response");
 	}
 
 	@Override
