+++ src/main/java/site/sonisori/sonisori/auth/oauth2/dto/OAuth2UserDto.java
@@ -1,9 +1,13 @@
 package site.sonisori.sonisori.auth.oauth2.dto;
 
+import site.sonisori.sonisori.common.enums.Role;
+import site.sonisori.sonisori.common.enums.SocialType;
+
 public record OAuth2UserDto(
 	String name,
 	String username,
-	String role,
-	String email
+	Role role,
+	String email,
+	SocialType socialType
 ) {
 }
