+++ src/main/java/site/sonisori/sonisori/repository/UserRepository.java
@@ -1,5 +1,7 @@
 package site.sonisori.sonisori.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -7,4 +9,5 @@ import site.sonisori.sonisori.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
+	Optional<User> findByUsername(String username);
 }
